requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
websockets==11.0.3
python-multipart==0.0.20
httpx==0.28.1
requests==2.32.3
paramiko==3.4.0
pydantic==2.5.0
alembic==1.13.1
pytest==7.4.3
pytest-asyncio==0.21.1
pydantic-settings